Fix type of tx and changing device_single back to a block size of 1024

+++ src/heuristicPlayout.cu
@@ -14,7 +14,7 @@
 #define BLOCK_SIZE 1024
 
 __global__ void heuristicPlayoutKernel(State *states, PlayerId *results, size_t numStates) {
-  uint8_t tx = threadIdx.x;
+  uint32_t tx = threadIdx.x;
   uint32_t bx = blockIdx.x;
   uint32_t tid = tx + (bx * NUM_LOCS);
